backend/requirements.txt
Flask>=3.0,<4.0
google-genai>=1.60.0,<2.0
python-dotenv>=1.0.0
